Rename orchestration/ to sia/ and add CLI entry point
- Rename orchestration/ → sia/ to match package name in pyproject.toml
- Add sia/__init__.py with version
- Add [project.scripts] so `pip install sia-agent` makes `sia` CLI
  available (e.g. `sia --task-dir ./tasks/spaceship-titanic --max_gen 5`)
- Update all internal imports to use sia.* package paths
- Update tests to import from sia.* directly (remove sys.path hacks)
- Update CI workflow to reference sia/ instead of orchestration/
