+++ .github/workflows/ci.yml
@@ -24,22 +24,19 @@ jobs:
       - name: Install uv
         uses: astral-sh/setup-uv@v4
 
-      - name: Install dependencies
+      - name: Install package
         run: |
-          uv pip install --system python-dotenv numpy pandas scikit-learn
+          uv pip install --system -e ".[dev]"
 
-      - name: Check Python syntax
+      - name: Verify CLI entry point
         run: |
-          python -m py_compile orchestration/orchestrator.py
-          python -m py_compile orchestration/context_manager.py
-          python -m py_compile orchestration/util.py
-          echo "All modules compile successfully"
+          sia --help
 
-      - name: Verify core imports
+      - name: Verify package imports
         run: |
-          cd orchestration
-          python -c "from context_manager import ContextManager; print('context_manager OK')"
-          python -c "import util; print('util OK')"
+          python -c "from sia.orchestrator import main, load_agent_execution; print('orchestrator OK')"
+          python -c "from sia.context_manager import ContextManager; print('context_manager OK')"
+          python -c "from sia.util import run_agent; print('util OK')"
 
       - name: Build package
         run: |
@@ -69,9 +66,9 @@ jobs:
       - name: Install uv
         uses: astral-sh/setup-uv@v4
 
-      - name: Install test dependencies
+      - name: Install package with test dependencies
         run: |
-          uv pip install --system pytest numpy pandas scikit-learn python-dotenv
+          uv pip install --system -e ".[dev]"
 
       - name: Run tests
         run: |
@@ -90,10 +87,10 @@ jobs:
           uv pip install --system ruff ty
 
       - name: Ruff check
-        run: ruff check orchestration/ tests/
+        run: ruff check sia/ tests/
 
       - name: Ruff format check
-        run: ruff format --check orchestration/ tests/
+        run: ruff format --check sia/ tests/
 
       - name: Type check with ty
-        run: ty check orchestration/
+        run: ty check sia/
