+++ sia/__init__.py
@@ -1,0 +1,3 @@
+"""SIA: Self-Improving Auto-researcher — an autonomous AI scientist framework."""
+
+__version__ = "0.1.1"
